server/requirements.txt
flask
flask-cors
psycopg2-binary
flask_jsonrpc
